Add astyle linting to Travis
This allows checking the code for style errors and leads to a nicer and more
consistent codebase.

Currently, most files are whitelisted in ci/lint-whitelist.txt, which prevents
the linter from checking them.  They can be removed from the whitelist once
they have been formatted properly.  Astyle output shows where style errors are
present, as well as how they should be fixed.

The aim is to eventually use astyle to lint all pilight source files, which is
is very useful for pull requests, in order to ensure a consistent codestyle.